YG-130 💡refactor: userRecommendation 로직 분리
